Move Menu to top-level components
