+++ electron/app/components/Menu.tsx
